requirements.txt
azure-functions==1.20.0
dnspython==2.6.1
pymongo==4.8.0
